% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atget.R
\name{atget}
\alias{atget}
\title{A Academic Torrent R download function}
\usage{
atget(name)
}
\arguments{
\item{To}{be added in the future}
}
\description{
This function allows you to download data through torrent.
}
\examples{
atget()
}
\keyword{download}
